I believe this is actually working

+++ src/model.ml
@@ -2,7 +2,7 @@ open Types
 
 type model =
   { part : part;
-    number : int;
+    number : float;
     production_map : production Production.t;
     production : production;
   }
